Share CLI connection logic to cli-tools
This will be used by xtask when using actions that require a connection like `applet-install`.
