+++ crates/cli-tools/src/lib.rs
@@ -17,6 +17,7 @@
 //! This library is also used for the internal maintenance CLI of Wasefire called xtask.
 
 #![feature(path_add_extension)]
+#![feature(try_find)]
 
 macro_rules! debug {
     ($($x:tt)*) => {
