+++ crates/cli/Cargo.toml
@@ -20,7 +20,6 @@ anyhow = { version = "1.0.86", default-features = false }
 clap = { version = "4.5.4", default-features = false, features = ["default", "derive", "env"] }
 clap_complete = { version = "4.5.2", default-features = false }
 data-encoding = { version = "2.6.0", default-features = false, features = ["std"] }
-humantime = { version = "2.1.0", default-features = false }
 rusb = { version = "0.9.4", default-features = false }
 wasefire-cli-tools = { version = "0.1.1-git", path = "../cli-tools" }
 wasefire-protocol = { version = "0.1.1-git", path = "../protocol", features = ["host"] }
